requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg # Ou outro driver de DB assíncrono que você use
pydantic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt] # <-- Biblioteca para hashing de senhas
python-multipart # Para upload de arquivos
httpx
bcrypt
wsgidav
icalendar
pytz
defusedxml
psycopg2
requests
alembic
python-dateutil
apscheduler
fastapi_mail